Server/requirements.txt
Django==4.2.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.4
gunicorn
dj-database-url==2.1.0
whitenoise==6.6.0
